requirements.txt
django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0

python-decouple==3.8
celery==5.4.0
redis==5.0.8
psycopg2-binary==2.9.9
drf-nested-routers==0.94.1
gunicorn==23.0.0
Pillow==10.4.0